Desenvolvimento da "Transação"

+++ BudgetBuddy.Domain/Entities/ContasBancarias/ContaBancaria.cs
@@ -3,7 +3,7 @@
     public class ContaBancaria : EntityBase
     {
         public string Nome { get; set; }
-        public double Saldo { get; set; }
+        public decimal Saldo { get; set; }
         public string Icon { get; set; }
         public int IdCategoria { get; set; }
     }
